random: give the RNG state a fixed 32-bit width for portable saves
The per-class RNG seeds were stored as unsigned long.  That type is 64-bit on
LP64 targets (Linux/macOS) and 32-bit on LLP64 (Windows/MinGW), so the seed
array had a different size on each, and the linear-congruential step ran at a
different width too.  A savegame's RNG block therefore could not be exchanged
between builds (the load over- or under-ran the block by 4 bytes per class),
and the generator could even diverge between platforms during play.

Store the seeds as uint32_t and run the LCG on a uint32_t accumulator, which is
the width the Boom generator was designed for.  The state now has the same
layout everywhere, savegames carry it identically across platforms, and the
defined 32-bit wraparound matches the original generator.  The vanilla demo
hash is unchanged (vanilla demo playback uses the lookup table, not the LCG).

+++ src/m_random.c
@@ -93,7 +93,7 @@ int (P_Random)(pr_class_t pr_class)
     (rng.prndindex = (rng.prndindex + 1) & 255) :
     (rng. rndindex = (rng. rndindex + 1) & 255) ;
 
-  unsigned long boom;
+  uint32_t boom;
 
   // killough 3/31/98:
   // If demo sync insurance is not requested, use
@@ -138,7 +138,7 @@ int (P_Random)(pr_class_t pr_class)
 void M_ClearRandom (void)
 {
   int i;
-  unsigned long seed = rngseed*2+1;    // add 3/26/98: add rngseed
+  uint32_t seed = rngseed*2+1;         // add 3/26/98: add rngseed
   for (i=0; i<NUMPRCLASS; i++)         // go through each pr_class and set
     rng.seed[i] = seed *= 69069ul;     // each starting seed differently
   rng.prndindex = rng.rndindex = 0;    // clear two compatibility indices
